server/requirements.txt
Flask==2.1.2
flask-marshmallow==0.14.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
marshmallow-sqlalchemy==0.28.0
psycopg2-binary==2.9.3
PyJWT==2.4.0
loguru==0.6.0
pika==1.2.1
